% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_div_survey_scheme.R
\name{env_div_survey_scheme}
\alias{env_div_survey_scheme}
\title{Environmental diversity survey scheme}
\usage{
env_div_survey_scheme(
  site_data,
  cost_column,
  survey_budget,
  env_vars_columns,
  method = "mahalanobis",
  locked_in_column = NULL,
  locked_out_column = NULL,
  exclude_locked_out = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{site_data}{\code{\link[sf:sf]{sf::sf()}} object containing the candidate survey
sites.}

\item{cost_column}{\code{character} name of the column in the argument to
the argument to \code{site_data} that contains the cost for surveying each
site. No missing (\code{NA}) values are permitted.}

\item{survey_budget}{\code{numeric} vector of maximum budgets for the survey
schemes. No missing (\code{NA}) values are permitted.}

\item{env_vars_columns}{\code{character} vector names of the columns in
the argument to \code{site_data} that contain \code{numeric} environmental
variables. No missing (\code{NA}) values are permitted.}

\item{method}{\code{character} name of the distance metric to use for
calculating environmental dissimilarity scores. See
\code{\link[vegan:vegdist]{vegan::vegdist()}} documentation the \code{method} parameter
for other available distance metrics and more information.
No missing (\code{NA}) values are permitted.
Defaults to \code{"mahalanobis"} for Mahalanobis distances.}

\item{locked_in_column}{\code{character} (optional) name of the column in
the argument to \code{site_data} that contains \code{logical}
(\code{TRUE}/ \code{FALSE}) values indicating if certain sites should be
locked into the survey scheme.
No missing (\code{NA}) values are permitted.
Defaults to \code{NULL} such that no sites are locked in.}

\item{locked_out_column}{\code{character} (optional) name of the column in
the argument to \code{site_data} that contains \code{logical}
(\code{TRUE}/ \code{FALSE}) values indicating if certain sites should be
locked out of the survey scheme.
No missing (\code{NA}) values are permitted.
Defaults to \code{NULL} such that no sites are locked out.}

\item{exclude_locked_out}{\code{logical} should locked out planning units
be entirely excluded from the optimization process?
Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} indicating if information should be
printed while generating survey scheme(s). Defaults to \code{FALSE}.}
}
\value{
\code{matrix} of \code{logical} (\code{TRUE}/ \code{FALSE})
values indicating if a site is selected in a scheme or not. Columns
correspond to sites, and rows correspond to different schemes.
}
\description{
Generate a survey scheme by maximizing the diversity of environmental
conditions that are surveyed.
}
\details{
The integer programming formulation of the environmental diversity
reserve selection problem (Faith & Walker 1996) is used to generate survey
schemes.
}
\examples{
# set seed for reproducibility
set.seed(123)

# simulate data
 x <- sf::st_as_sf(
   tibble::tibble(x = rnorm(4), y = rnorm(4),
                  v1 = c(0.1, 0.2, 0.3, 10), # environmental axis 1
                  v2 = c(0.1, 0.2, 0.3, 10), # environmental axis 2
                  cost = rep(1, 4)),
    coords = c("x", "y"))

# plot the sites' environmental conditions
plot(x[, c("v1", "v2")], pch = 16, cex = 3)

# generate scheme with a budget of 2
s <- env_div_survey_scheme(x, "cost", 2, c("v1", "v2"), "mahalanobis")

# print scheme
print(s)

# plot scheme
x$scheme <- c(s)
plot(x[, "scheme"], pch = 16, cex = 3)
}
\references{
Faith DP & Walker PA (1996) Environmental diversity: on the best-possible
use of surrogate data for assessing the relative biodiversity of sets of
areas. \emph{Biodiversity & Conservation}, \strong{5}, 399--415.
}
