% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evdci.R
\name{evdci}
\alias{evdci}
\title{Expected value of the decision given current information}
\usage{
evdci(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_management_cost_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  total_budget,
  site_management_locked_in_column = NULL,
  site_management_locked_out_column = NULL,
  prior_matrix = NULL,
  optimality_gap = 0
)
}
\arguments{
\item{site_data}{\code{\link[sf]{sf}} object with site data.}

\item{feature_data}{\code{\link[base]{data.frame}} object with feature data.}

\item{site_occupancy_columns}{\code{character} names of \code{numeric}
columns in the argument to \code{site_data} that contain detections
and non-detections of features at each site.
Each column should correspond to a different feature, and contain
binary detection/non-detection value (zeros or ones) indicating if the
feature was detected in a previous survey or not. If a site has not
been surveyed before, then missing (\code{NA}) values should be used.
Additionally, if a feature was not looked for when surveying a specific
site, then missing (\code{NA}) value should be used.}

\item{site_probability_columns}{\code{character} names of \code{numeric}
columns in the argument to \code{site_data} that contain modelled
probabilities of occupancy for each feature in each site.
Each column should correspond to a different feature, and contain
probability data (values between zero and one). No missing (\code{NA})
values are permitted in these columns.}

\item{site_management_cost_column}{\code{character} name of column in the
argument to \code{site_data} that contains costs for managing each
site for conservation. This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_survey_sensitivity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of future surveys correctly detecting a presence of each
feature in a given site (i.e. the sensitivity of the survey methodology).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.}

\item{feature_survey_specificity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of future surveys correctly detecting an absence of each
feature in a given site (i.e. the specificity of the survey methodology).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.}

\item{feature_model_sensitivity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of the initial models correctly predicting a presence of each
feature in a given site (i.e. the sensitivity of the models).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.
This should ideally be calculated using \code{\link{fit_occupancy_models}}.}

\item{feature_model_specificity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of the initial models correctly predicting an absence of each
feature in a given site (i.e. the specificity of the models).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.
This should ideally be calculated using \code{\link{fit_occupancy_models}}.}

\item{feature_preweight_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{preweight}
values used to parametrize  the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_postweight_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{postweight}
values used to parametrize  the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_target_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{target}
values used to parametrize the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{total_budget}{\code{numeric} maximum expenditure permitted
for conducting surveys and managing sites for conservation.}

\item{site_management_locked_in_column}{\code{character} name of the column
in the argument to \code{site_data} that contains \code{logical}
(\code{TRUE} / \code{FALSE}) values indicating which sites should
be locked in for (\code{TRUE}) being managed for conservation or
(\code{FALSE}) not. No missing (\code{NA}) values are permitted in this
column. This is useful if some sites have already been earmarked for
conservation, or if some sites are already being managed for conservation.
Defaults to \code{NULL} such that no sites are locked in.}

\item{site_management_locked_out_column}{\code{character} name of the column
in the argument to \code{site_data} that contains \code{logical}
(\code{TRUE} / \code{FALSE}) values indicating which sites should
be locked out for (\code{TRUE}) being managed for conservation or
(\code{FALSE}) not. No missing (\code{NA}) values are permitted in this
column. This is useful if some sites could potentially be surveyed
to improve model predictions even if they cannot be managed for
conservation. Defaults to \code{NULL} such that no sites are locked out.}

\item{prior_matrix}{\code{numeric} \code{matrix} containing
the prior probability of each feature occupying each site.
Rows correspond to features, and columns correspond to sites.
Defaults to \code{NULL} such that prior data is calculated automatically
using \code{\link{prior_probability_matrix}}.}

\item{optimality_gap}{\code{numeric} relative optimality gap for generating
conservation prioritizations. A value of zero indicates that
prioritizations must be solved to optimality. A value of 0.1 indicates
prioritizations must be within 10\% of optimality. Defaults to 0.}
}
\value{
\code{numeric} value.
}
\description{
Calculate the \emph{expected value of the conservation management decision
given current information}. This metric describes the value of the management
decision that is expected when the decision maker is limited to
existing biodiversity data (i.e. survey data and environmental niche models).
}
\details{
This function calculates the expected value and does not
 use approximation methods. As such, this function can only be applied
 to very small problems.
}
\examples{
# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)
feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)
print(feature_data)

# set total budget for managing sites for conservation
total_budget <- sum(site_data$management_cost) * 0.5

# calculate expected value of management decision given current information
# using exact method
ev_current <- evdci(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  "management_cost", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "preweight", "postweight", "target", total_budget)

# print exact value
print(ev_current)

}
\seealso{
\code{\link{prior_probability_matrix}}.
}
