% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_occupancy_models.R
\name{fit_occupancy_models}
\alias{fit_occupancy_models}
\title{Fit models to estimate probability of occupancy}
\usage{
fit_occupancy_models(
  site_data,
  site_occupancy_columns,
  site_env_vars_columns,
  parameters,
  n_folds = rep(ifelse(nrow(site_data) > 1000, 10, 5), length(site_occupancy_columns)),
  site_weight_columns = NULL,
  n_threads = 1,
  verbose = FALSE
)
}
\arguments{
\item{site_data}{\code{\link[sf]{sf}} object with site data.}

\item{site_occupancy_columns}{\code{character} names of \code{numeric}
columns in the
argument to \code{site_data} that contain presence/absence data.
Each column should correspond to a different feature, and contain
binary presence/absence data (zeros or ones) indicating if the
feature was detected in a previous survey or not. If a site has not
been surveyed before, then missing (\code{NA}) values should be used.}

\item{site_env_vars_columns}{\code{character} names of columns in the
argument to \code{site_data} that contain environmental information
for fitting updated occupancy models based on possible survey outcomes.
Each column should correspond to a different environmental variable,
and contain \code{numeric}, \code{factor}, or \code{character} data.
No missing (\code{NA}) values are permitted in these columns.}

\item{parameters}{\code{list} of \code{list} objects containing the
parameters for fitting models for each
feature. Ideally, these parameters would be determined using the
\code{\link{tune_occupancy_models}} function. Note that arguments must
have a \code{nrounds} element (see example below).
Valid parameters include: \code{"max_depth"}, \code{"eta"},
\code{"nrounds"}, \code{"lambda"}, \code{"subsample"},
\code{"colsample_bytree"}, and \code{"objective"}.
See documentation for the \code{params} argument in
\code{\link[xgboost]{xgb.train}} for more information.}

\item{n_folds}{\code{numeric} number of folds to split the training
data into when fitting models for each feature.
Defaults to 10 if there are more than 1000
rows in the argument \code{site_data}, otherwise defaults to 5.
Note that if the maximum number of minority cases (e.g. 1 presence and
1000 absences) is fewer than the number of folds then one minority case
is sampled per fold.}

\item{site_weight_columns}{\code{character} name of columns in
\code{site_data} containing weights for model fitting. These columns must
contain \code{numeric} values. No missing (\code{NA}) values are
permitted. Defaults to \code{NULL} such that all data are given
equal weight when fitting models.}

\item{n_threads}{\code{integer} number of threads to use for fitting models.
Defaults to 1.}

\item{verbose}{\code{logical} indicating if information should be
printed during computations. Defaults to \code{FALSE}.}
}
\value{
\code{list} object containing the following elements:
 \describe{
 \item{predictions}{\code{\link[tibble]{tibble}} with estimated
   probabilities of each feature occupying each site. Rows
   correspond to sites and columns correspond to features.}
 \item{performance}{\code{\link[tibble]{tibble}} with model information.
   Each row corresponds to a different feature and column contain
   different information. Specifically, the columns contain:
   (\code{name}) \code{character} name of the features,
   (\code{sensitivity}) \code{numeric} model sensitivities,
   (\code{specificity}) \code{numeric} model specificities, and
   (\code{auc}) \code{numeric} model area under the curve (AUC) statistics.}
 }
}
\description{
Estimate probability of occupancy for a set of features in a set of
planning units.
}
\details{
The models are fitted using the specified tuning parameters and
 automatically calibrated \code{scale_pos_weight}
 \code{\link[xgboost]{xgboost}} parameters
 (i.e. number of absence divided by number of presences per feature)
 to account for unbalanced datasets. After the models are fitted, they
 are used to make probability of occupancy predictions (via
 \code{\link[xgboost]{xgboost}}). The average sensitivity, specificity,
 and area under the curve (AUC) statistics are calculated using the data held
 are then reported to assess model performance. Note that each feature
 is run with the same seed to ensure reproducibility.
}
\examples{
# set seed
set.seed(100)

# simulate data for 200 sites, 2 features, and 3 environmental variables
 x <- simulate_site_data(200, 2, 0.5, n_env_vars = 3)

# set parameters manually
xgb_params <- list(list(nrounds = 10, objective = "binary:logistic"),
                   list(nrounds = 5, objective = "binary:logistic"))

# fit models
m <- fit_occupancy_models(x, c("f1", "f2"), c("e1", "e2", "e3"),
                          xgb_params)

# preview models predictions
print(m$predictions)

# preview models performance
print(m$performance)

}
