% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_states.R
\name{n_states}
\alias{n_states}
\title{Number of states}
\usage{
n_states(n_sites, n_features)
}
\arguments{
\item{n_sites}{\code{numeric} number of sites.}

\item{n_features}{\code{numeric} number of features.}
}
\value{
A \code{numeric} value.
}
\description{
Calculate the total number of presence/absence states for a given
number of sites and features.
}
\examples{
# calculate number of states for 3 sites and 2 features
n_states(n_sites = 2, n_features = 3)

}
