% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conservation_value.R
\name{plot_conservation_value}
\alias{plot_conservation_value}
\title{Plot conservation value}
\usage{
plot_conservation_value(
  site_data,
  feature_data,
  site_occupancy_columns,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  plot = TRUE
)
}
\arguments{
\item{site_data}{\code{\link[sf]{sf}} object with site data.}

\item{feature_data}{\code{\link[base]{data.frame}} object with feature data.}

\item{site_occupancy_columns}{\code{character} names of \code{numeric}
columns in the
argument to \code{site_data} that contain presence/absence data.
Each column should correspond to a different feature, and contain
binary presence/absence data (zeros or ones) indicating if the
feature was detected in a previous survey or not. If a site has not
been surveyed before, then missing (\code{NA}) values should be used.}

\item{feature_preweight_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{preweight}
values used to parametrize  the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_postweight_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{postweight}
values used to parametrize  the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_target_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{target}
values used to parametrize the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{plot}{\code{logical} should the plot be returned? If \code{FALSE}
the underlying data (i.e. a \code{tibble}{tibble} object) is returned
instead of the plot. Defaults to \code{TRUE}.}
}
\value{
\code{\link[ggplot2]{ggplot}} plot object.
}
\description{
Create a plot showing the relative conservation value for different
conservation features.
}
\details{
This function creates a plot to visualize the conservation value
  for each feature. The x-axis shows the number of protected sites that
  are occupied by a given feature, and the y-axis represents the amount of
  conservation value with that number of protected sites. Lines correspond
  to features. Each feature will have a line that is comprised of a
  solid component and a dashed component. The solid component of the line
  corresponds to the range of values based on recorded
  presences and absences in the site data. The dashed component of the
  line corresponds to the additional values that could
  be obtained if additional surveys were conducted and the feature was
  detected in all sites that are currently unsurveyed. Features that are
  associated with greater values (greater y-axis
  values) at lower numbers of protected sites (smaller x-axis values) will,
  broadly speaking, be considered more important for protection.
}
\examples{
# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 4, prop = 0.5)
feature_data <- simulate_feature_data(n_features = 4, prop = 0.5)

# preview simulated data
print(site_data)
print(feature_data)

# plot conservation value for each feature
plot_conservation_value(
 site_data = site_data,
 feature_data = feature_data,
 site_occupancy_columns = paste0("f", seq_len(4)),
 feature_preweight_column = "preweight",
 feature_postweight_column  = "postweight",
 feature_target_column = "target")

}
