% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_probability_matrix.R
\name{prior_probability_matrix}
\alias{prior_probability_matrix}
\title{Prior probability matrix}
\usage{
prior_probability_matrix(site_data, feature_data, site_occupancy_columns,
  site_probability_columns, feature_survey_sensitivity_column,
  feature_survey_specificity_column, feature_model_sensitivity_column)
}
\value{
\code{matrix} object containing the prior probabilities of each
  feature occupying each site. Each row corresponds to a different
  feature and each column corresponds to a different site.
}
\description{
Create prior probability matrix for the value of information analysis.
}
\details{
The prior matrix is constructed using a combination of previous
  survey results and modelled predictions for sites that have not been
  surveyed. Let \eqn{I} denote the set of features (indexed by \eqn{i}) and
  let \eqn{J} denote the set of sites (indexed by \eqn{j}).
  Let \eqn{D_j} indicate which sites have already been surveyed (using zeros
  and ones).
  Let \eqn{S_i} and \eqn{N_i} denote sensitivity and specificity
  of the survey method for features \eqn{i \in I} (respectively).
  Also let \eqn{H_{ij}} denote the presence or absence of each feature
  in each site if they have been surveyed (using zeros or ones),
  with unsurveyed sites denoted with -1 values
  Let \eqn{{H'}_{ij}} indicate the modelled probability of each feature
  occupying each site. Finally, let \eqn{{S'}_i} and \eqn{{N'}_i} denote
  sensitivity and specificity of the model for feature \eqn{i \in I}.
  Given such data, the (\eqn{P_{ij}}) prior probability of feature \eqn{i}
  occupying site \eqn{j} is:

\deqn{
P_{ij} = \\
S_i, \text{ if } D_j = 1, H_{ij} = 1 \space (i \text{ detected in } j) \\
1 - N_i, \text{ else if } D_j = 1, H_{ij} = 0 \space (i \text{ not detected in } j) \\
{S'}_i {H'}_{ij}, \text{ else if } D_j = 0 \space (j \text{ not surveyed)} \\
}
}
\examples{
# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2,
                                prop = 0.5)
feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# calculate prior probability matrix
prior_probability_matrix(site_data, feature_data, c("f1", "f2"),
                         c("p1", "p2"), "sensitivity", "specificity")

}
