% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\alias{sim_sites}
\alias{sim_features}
\title{Simulated datasets}
\format{
TODO.
}
\usage{
data(sim_features)

data(sim_sites)
}
\description{
Simulated data for prioritizing sites for ecological surveys.
}
\details{
TODO.
}
\examples{
# load data
data(sim_sites, sim_features)

# print feature data
print(sim_features, width = Inf)
# print site data
print(sim_sites, width = Inf)

}
\keyword{datasets}
