% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_evdpi.R
\name{approx_evdpi}
\alias{approx_evdpi}
\title{Approximate expected value of decision given perfect information}
\usage{
approx_evdpi(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_management_cost_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  total_budget,
  site_management_locked_in_column = NULL,
  prior_matrix = NULL,
  n_approx_obj_fun_points = 1000,
  optimality_gap = 0,
  n_approx_replicates = 100,
  n_approx_states_per_replicate = 1000,
  method_approx_states = "weighted_without_replacement",
  seed = 500
)
}
\arguments{
\item{site_data}{\code{\link[sf]{sf}} object with site data.}

\item{feature_data}{\code{\link[base]{data.frame}} object with feature data.}

\item{site_occupancy_columns}{\code{character} names of \code{numeric}
columns in the
argument to \code{site_data} that contain presence/absence data.
Each column should correspond to a different feature, and contain
binary presence/absence data (zeros or ones) indicating if the
feature was detected in a previous survey or not. If a site has not
been surveyed before, then missing (\code{NA}) values should be used.}

\item{site_probability_columns}{\code{character} names of \code{numeric}
columns in the argument to \code{site_data} that contain modelled
probabilities of occupancy for each feature in each site.
Each column should correspond to a different feature, and contain
probability data (values between zero and one). No missing (\code{NA})
values are permitted in these columns.}

\item{site_management_cost_column}{\code{character} name of column in the
argument to \code{site_data} that contains costs for managing each
site for conservation. This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_survey_sensitivity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of future surveys correctly detecting a presence of each
feature in a given site (i.e. the sensitivity of the survey methodology).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.}

\item{feature_survey_specificity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of future surveys correctly detecting an absence of each
feature in a given site (i.e. the specificity of the survey methodology).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.}

\item{feature_model_sensitivity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of the initial models correctly predicting a presence of each
feature in a given site (i.e. the sensitivity of the models).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.
This should ideally be calculated using \code{\link{fit_occupancy_models}}.}

\item{feature_model_specificity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of the initial models correctly predicting an absence of each
feature in a given site (i.e. the specificity of the models).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.
This should ideally be calculated using \code{\link{fit_occupancy_models}}.}

\item{feature_preweight_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{preweight}
values used to parametrize  the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_postweight_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{postweight}
values used to parametrize  the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_target_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{target}
values used to parametrize the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{total_budget}{\code{numeric} maximum expenditure permitted
for conducting surveys and managing sites for conservation.}

\item{site_management_locked_in_column}{\code{character} name of the column
in the argument to \code{site_data} that contains \code{logical}
(\code{TRUE} / \code{FALSE}) values indicating which sites should
be locked in for (\code{TRUE}) being managed for conservation or
(\code{FALSE}) not. No missing (\code{NA}) values are permitted in this
column. This is useful if some sites have already been earmarked for
conservation, or if some sites are already being managed for conservation.
Defaults to \code{NULL} such that no sites are locked in.}

\item{prior_matrix}{\code{numeric} \code{matrix} containing
the prior probability of each feature occupying each site.
Rows correspond to features, and columns correspond to sites.
Defaults to \code{NULL} such that prior data is calculated automatically
using \code{\link{prior_probability_matrix}}.}

\item{n_approx_obj_fun_points}{\code{integer} number of points to use
for approximating the piecewise-linear components of the objective
function. Greater values result in more precise calculations, but
also incur more computational costs. Defaults to 1000.}

\item{optimality_gap}{\code{numeric} relative optimality gap for generating
conservation prioritizations. A value of zero indicates that
prioritizations must be solved to optimality. A value of 0.1 indicates
prioritizations must be within 10\% of optimality. Defaults to 0.}

\item{n_approx_replicates}{\code{integer} number of replicates to use for
approximating the expected value calculations. Defaults to 100.}

\item{n_approx_states_per_replicate}{\code{integer} number of states to use
per replicate for approximating the expected value of a given management
action. This number must be smaller than or equal to the total number of
presence absence states in the system
(i.e. \code{n_states(nrow(site_data), nrow(feature_data))})
Defaults to 1000.}

\item{method_approx_states}{\code{character} name of method that is
used to sample states for approximating the expected value
calculations. Available options are:
\code{"uniform_with_replacement"}, \code{"uniform_without_replacement"},
\code{"weighted_with_replacement"}, \code{"weighted_without_replacement"}.
Uniform sampling methods have an equal chance of returning each
state, and weighted sampling methods are more likely to return
states with a higher prior probability of occurring.
Defaults to \code{"weighted_without_replacement"}.}

\item{seed}{\code{integer} random seed used for generating states. Defaults
to 500.}
}
\value{
\code{numeric} \code{vector} containing the estimate for each
  replicate.
}
\description{
Calculate the \emph{expected value of the conservation management decision
given perfect information}. This metric describes the value of the management
decision that is expected when the decision maker is omniscient (i.e. they
know exactly which features occur in which sites). Although real-world
conservation planing scenarios rarely involve such omniscient decision
makers, this metric is useful to provide an upper bound on the expected
value of management decisions following additional data collection.
}
\examples{
# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)
feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)
print(feature_data)

# set total budget for managing sites for conservation
total_budget <- sum(site_data$management_cost) * 0.5

# calculate expected value of management decision given perfect information
# using approximate method with 100 replicates and 50 states per replicate
ev_prime_certainty <- approx_evdci(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  "management_cost", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "preweight", "postweight", "target",
  total_budget, n_approx_replicates = 100,
  n_approx_states_per_replicate = 50)

# print approximate value
print(ev_prime_certainty)

}
\seealso{
\code{\link{prior_probability_matrix}},
\code{\link{evdpi}}.
}
