% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_evdsi.R
\name{approx_evdsi}
\alias{approx_evdsi}
\title{Expected value of the decision given survey information}
\usage{
approx_evdsi(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_env_vars_columns,
  site_management_cost_column,
  site_survey_scheme_column,
  site_survey_cost_column,
  feature_survey_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_alpha_column,
  feature_gamma_column,
  total_budget,
  xgb_parameters,
  site_management_locked_in_column = NULL,
  prior_matrix = NULL,
  n_approx_obj_fun_points = 1000,
  optimality_gap = 0,
  site_weight_columns = NULL,
  xgb_n_folds = rep(5, nrow(feature_data)),
  n_approx_replicates = 100,
  n_approx_states_per_replicate = 1000,
  seed = 500
)
}
\arguments{
\item{site_data}{\code{\link[sf]{sf}} object with site data.}

\item{feature_data}{\code{\link[base]{data.frame}} object with feature data.}

\item{site_occupancy_columns}{\code{character} names of \code{numeric}
columns in the
argument to \code{site_data} that contain presence/absence data.
Each column should correspond to a different feature, and contain
binary presence/absence data (zeros or ones) indicating if the
feature was detected in a previous survey or not. If a site has not
been surveyed before, then missing (\code{NA}) values should be used.}

\item{site_probability_columns}{\code{character} names of \code{numeric}
columns in the argument to \code{site_data} that contain modelled
probabilities of occupancy for each feature in each site.
Each column should correspond to a different feature, and contain
probability data (values between zero and one). No missing (\code{NA})
values are permitted in these columns.}

\item{site_env_vars_columns}{\code{character} names of columns in the
argument to  \code{site_data} that contain environmental information
for fitting updated occupancy models based on possible survey outcomes.
Each column should correspond to a different environmental variable,
and contain \code{numeric}, \code{factor}, or \code{character} data.
No missing (\code{NA}) values are permitted in these columns.}

\item{site_management_cost_column}{\code{character} name of column in the
argument to \code{site_data} that contains costs for managing each
site for conservation. This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{site_survey_scheme_column}{\code{character} name of \code{logical}
(\code{TRUE} / \code{FALSE}) column in the argument to \code{site_data}
that indicates which sites are selected in the scheme or not.
No missing \code{NA} values are permitted. Additionally, only sites
that are missing data can be selected or surveying (as per the
argument to \code{site_occupancy_columns}).}

\item{site_survey_cost_column}{\code{character} name of column in the
argument to  \code{site_data} that contains costs for surveying each
site. This column should have \code{numeric} values that are equal to
or greater than zero. No missing (\code{NA}) values are permitted in this
column.}

\item{feature_survey_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains \code{logical} (\code{TRUE} /
\code{FALSE}) values indicating if the feature will be surveyed in
the planned surveys or not. Note that considering additional features will
rapidly increase computational burden, and so it is only recommended to
consider features that are of specific conservation interest.
No missing (\code{NA}) values are permitted in this column.}

\item{feature_survey_sensitivity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of future surveys correctly detecting a presence of each
feature in a given site (i.e. the sensitivity of the survey methodology).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.}

\item{feature_survey_specificity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of future surveys correctly detecting an absence of each
feature in a given site (i.e. the specificity of the survey methodology).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.}

\item{feature_model_sensitivity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of the initial models correctly predicting a presence of each
feature in a given site (i.e. the sensitivity of the models).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.
This should ideally be calculated using \code{\link{fit_occupancy_models}}.}

\item{feature_model_specificity_column}{\code{character} name of the
column in the argument to \code{feature_data} that contains
probability of the initial models correctly predicting an absence of each
feature in a given site (i.e. the specificity of the models).
This column should have \code{numeric} values that are between zero and
one. No missing (\code{NA}) values are permitted in this column.
This should ideally be calculated using \code{\link{fit_occupancy_models}}.}

\item{feature_alpha_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{\alpha}
values used to parametrize
the conservation benefit of managing of each feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{feature_gamma_column}{\code{character} name of the column in the
argument to \code{feature_data} that contains the \eqn{\gamma}
values used to parametrize the conservation benefit of managing of each
feature.
This column should have \code{numeric} values that
are equal to or greater than zero. No missing (\code{NA}) values are
permitted in this column.}

\item{total_budget}{\code{numeric} maximum expenditure permitted
for conducting surveys and managing sites for conservation.}

\item{xgb_parameters}{\code{list} of \code{list} objects
containing the parameters for fitting models for each
feature. See documentation for the \code{params} argument in
\code{\link[xgboost]{xgb.train}} for available parameters. Ideally,
these parameters would be determined using the
\code{\link{tune_occupancy_models}} function. Note that arguments must
have a \code{nrounds} element (see example below).}

\item{site_management_locked_in_column}{\code{character} name of the column
in the argument to \code{site_data} that contains \code{logical}
(\code{TRUE} / \code{FALSE}) values indicating which sites should
be locked in for (\code{TRUE}) being managed for conservation or
(\code{FALSE}) not. No missing (\code{NA}) values are permitted in this
column. This is useful if some sites have already been earmarked for
conservation, or if some sites are already being managed for conservation.
Defaults to \code{NULL} such that no sites are locked in.}

\item{prior_matrix}{\code{numeric} \code{matrix} containing
the prior probability of each feature occupying each site.
Rows correspond to features, and columns correspond to sites.
Defaults to \code{NULL} such that prior data is calculated automatically
using \code{\link{prior_probability_matrix}}.}

\item{n_approx_obj_fun_points}{\code{integer} number of points to use
for approximating the piecewise-linear components of the objective
function. Greater values result in more precise calculations, but
also incur more computational costs. Defaults to 1000.}

\item{optimality_gap}{\code{numeric} relative optimality gap for generating
conservation prioritizations. A value of zero indicates that
prioritizations must be solved to optimality. A value of 0.1 indicates
prioritizations must be within 10\% of optimality. Defaults to 0.}

\item{site_weight_columns}{\code{character} name of columns in
\code{site_data} containing weights for model fitting for each
feature. These columns must contain \code{numeric} values greater
than or equal to zero. No missing (\code{NA}) values are
permitted. Defaults to \code{NULL} such that all data are given
equal weight when fitting models.}

\item{xgb_n_folds}{\code{integer} vector containing the number of
 k-fold cross-validation folds to use for fitting models and
 assessing model performance for each feature. Ideally, the number of folds
 should be exactly the same as the number used for tuning the
 model parameters (i.e. same parameter to the \code{n_folds}
 argument in \link{tune_occupancy_models} when generating parameters
for \code{xgb_parameters}).}

\item{n_approx_replicates}{\code{integer} number of replicates to use for
approximating the expected value calculations. Defaults to 100.}

\item{n_approx_states_per_replicate}{\code{integer} number of states to use
per replicate for approximating the expected value of a given management
action. This number must be smaller than or equal to the total number of
presence absence states in the system
(i.e. \code{n_states(nrow(site_data), nrow(feature_data))})
Defaults to 1000.}

\item{seed}{\code{integer} state of the random number generator for
partitioning data into folds cross-validation and fitting \pkg{xgboost}
models. It is also used for generating states.
This parameter must remain the same to compare results from different
functions using the approximation methods. Defaults to 500.}
}
\value{
\code{numeric} \code{vector} containing the estimate for each
  replicate.
}
\description{
Calculate the \emph{expected value of the conservation management decision
given survey information}. This metric describes the value of the management
decision that is expected when the decision maker conducts a surveys a
set of sites to inform the decision.
}
\details{
This function uses approximation methods to estimate the
  expected value calculations. As such, you will need to ensure that
  the same seed is used when comparing results to other functions that
  use approximation methods. Additionally, the accuracy of these
  calculations depend on the arguments to
  \code{n_approx_replicates} and \code{n_approx_states_per_replicate}, and
  so you may need to increase these parameters for large problems.
}
\examples{
# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 15, n_features = 2, prop = 0.5)
feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)
print(feature_data)

# set total budget for managing sites for conservation
 # (i.e. 50\% of the cost of managing all sites)
total_budget <- sum(site_data$management_cost) * 0.5

# create a survey scheme that samples the first two sites that
# are missing data
site_data$survey_site <- FALSE
site_data$survey_site[which(is.na(site_data$f1))[1:2]] <- TRUE

# define xgboost tuning parameters
xgb_parameters <-
 list(list(objective = "binary:logistic", nrounds = 8, eta = 0.1))[rep(1, 2)]

# calculate expected value of management decision given the survey
# information using exact method
approx_ev_survey <- approx_evdsi(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  c("e1", "e2", "e3"), "management_cost", "survey_site",
  "survey_cost", "survey", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "alpha", "gamma", total_budget, xgb_parameters)

# print exact value
print(approx_ev_survey)

}
\seealso{
\code{\link{prior_probability_matrix}}.
}
