% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_feature_data.R
\name{simulate_feature_data}
\alias{simulate_feature_data}
\title{Simulate feature data}
\usage{
simulate_feature_data(n_features, proportion_of_survey_features)
}
\arguments{
\item{n_features}{\code{integer} number of features.}

\item{proportion_of_survey_features}{\code{numeric} proportion of features
that will be examined in the new surveys. Values must be between
zero and one.}
}
\value{
\code{\link[tibble]{tibble}} object. It contains the following
data:
\describe{
\item{\code{name}}{\code{character} name of each feature.}

\item{\code{survey}}{\code{logical} (\code{TRUE} / \code{FALSE}) values
  indicating if each feature should be examined in surveys or not.}

\item{\code{sensitivity}}{\code{numeric} sensitivity (true positive rate)
  of the survey methodology for each features.}

\item{\code{specificity}}{\code{numeric} specificity (true absence rate)
  of the survey methodology for each features.}

\item{\code{alpha}}{\code{numeric} values used to parametrize
  the conservation benefit of managing of each feature. See
  \code{\link{expected_value_of_survey_scheme}} documentation for
  more information.}

\item{\code{gamma}}{\code{numeric} values used to parametrize
  the conservation benefit of managing of each feature. See
  \code{\link{expected_value_of_survey_scheme}} documentation for
  more information.}

}
}
\description{
Simulate feature data for developing simulated survey schemes.
}
\examples{
# set seed for reproducibility
set.seed(123)

# simulate data
d <- simulate_feature_data(n_features = 5,
                           proportion_of_survey_features = 0.5)
# print data
print(d)

}
\seealso{
\code{\link{simulate_site_data}}
}
